rust: move rust tests to integration tests
