+++ Cargo.toml
@@ -18,6 +18,10 @@ name = "smart_keymap"
 crate-type = ["lib"]
 
 [[test]]
+name = "rust-integration"
+path = "tests/rust/keymap.rs"
+
+[[test]]
 name = "cucumber-keymap"
 path = "tests/cucumber/keymap.rs"
 harness = false
